Extract utility methods in JDBC module
